Fixed delayed lag issue (finally!). The client previously sent an extra ping packet every 500ms, but the robot only handled one packet per loop. Now only one packet is sent per loop.
Added send times to all packets.
Added latency display to printData().
Fixed an issue with robot and server occasionally calling printData() on ping packets,
Fixed a potential issue with time variable size in telepresence.cpp.

+++ telepresence.cpp
@@ -6,9 +6,9 @@
 #include <unistd.h>
 #include <string.h>
 #include <signal.h>
+#include <chrono>
 #include <exception>
 #include <stdexcept>
-#include <chrono>
 #include "server.h"
 #include "client.h"
 #include "robot.h"
@@ -23,7 +23,7 @@ bool running;
 //Functions
 int main(int argc, char *argv[]) {
 	running = false;
-	long elapsedTime = 0;
+	long long elapsedTime = 0;
 	std::chrono::steady_clock::time_point startTime = std::chrono::steady_clock::now();
 	bool isClient = false;
 	bool isRobot = false;
